requirements.txt
requests==2.31.0
uvicorn==0.22.0
fastapi==0.101.1
xmltodict==0.13.0
python-dotenv==1.0.1
python-multipart==0.0.9
pydantic==2.10.6